SA-832: Pulled out colors to constant
