+++ src/components/percentWheel/PercentWheel.module.scss
@@ -1,3 +1,5 @@
+@import '~@sparkpost/matchbox/src/styles/config.scss';
+
 .PercentWheel {
   position: relative;
