+++ src/pages/inboxPlacement/constants/graphs.js
@@ -1,0 +1,15 @@
+export const PERCENT_WHEEL = {
+  color: '#0CBAC7'
+};
+
+export const FOLDER_PLACEMENT_BAR_CHART = {
+  inbox: {
+    color: '#50D0D9'
+  },
+  spam: {
+    color: '#0CBAC7'
+  },
+  missing: {
+    color: '#00838C'
+  }
+};
