--- src/pages/inboxPlacement/constants/percentWheel.js
@@ -1,3 +1,0 @@
-export const PERCENT_WHEEL = {
-  color: '#0CBAC7'
-};
